+++ java-client/src/test/java/org/opensearch/client/opensearch/integTest/httpclient5/TasksIT.java
@@ -1,0 +1,5 @@
+package org.opensearch.client.opensearch.integTest.httpclient5;
+
+import org.opensearch.client.opensearch.integTest.AbstractTasksIT;
+
+public class TasksIT extends AbstractTasksIT implements HttpClient5TransportSupport {}